+++ generators/app/templates/grunt/config/jsMin.js
@@ -1,8 +1,0 @@
-/**
- <%- include('../../../../common/partial/header.txt') %>
- */
-/* eslint object-curly-newline:"off" */
-module.exports = {
-	dev    : {},
-	release: {}
-};