+++ generators/app/templates/grunt/config/usemin.js
@@ -1,8 +1,0 @@
-/**
- <%- include('../../../../common/partial/header.txt') %>
- */
-module.exports = {
-	html: [
-		'<%%= currentTarget %%>/index.html'
-	]
-};