+++ generators/app/writing/app.js
@@ -13,13 +13,13 @@
 	const utils = require('../../common/utils.js');
 
 	module.exports = $that => {
-		utils.copyFileTpl($that, 'app/app.js', {
+		utils.copyFileTpl($that, 'app/app.txt', {
 			appNameKebab: $that.appNameKebab,
 			appNameCamel: $that.appNameCamel,
 			authorShort : $that.authorShort,
 			nowDate     : $that.nowDate,
 			nowTime     : $that.nowTime
-		});
+		}, 'app/app.js');
 	};
 
 })();