+++ generators/app/writing/providers.js
@@ -28,13 +28,13 @@
 
 	module.exports = $that => {
 		_.forEach(providers, $provider => {
-			utils.copyFileTpl($that, 'app/config/providers/' + $provider + '.config.js', {
+			utils.copyFileTpl($that, 'app/config/providers/' + $provider + '.config.txt', {
 				appNameKebab: $that.appNameKebab,
 				appNameCamel: $that.appNameCamel,
 				authorShort : $that.authorShort,
 				nowDate     : $that.nowDate,
 				nowTime     : $that.nowTime
-			});
+			}, 'app/config/providers/' + $provider + '.config.js');
 		});
 	};