+++ generators/view/writing/jsFiles.js
@@ -31,8 +31,8 @@
 			stateAbstract  : $that.stateAbstract
 		};
 
-		utils.copyFileTpl($that, 'controller.txt', config, filePrefix + 'controller.js');
-		utils.copyFileTpl($that, 'service.txt', config, filePrefix + 'service.js');
+		utils.copyFileTplFrom($that, 'controller.txt', config, filePrefix + 'controller.js');
+		utils.copyFileTplFrom($that, 'service.txt', config, filePrefix + 'service.js');
 		utils.copyFileTpl($that, 'constant.txt', config, filePrefix + 'constant.js');
 		utils.copyFileTpl($that, 'state.txt', config, filePrefix + 'state.js');
 	};